Make code buffer units 8bit rather than 16bit.

+++ src/compiled_method.h
@@ -37,7 +37,7 @@ class CompiledMethod {
 #else
   // Create a CompiledMethod from the oatCompileMethod
   CompiledMethod(InstructionSet instruction_set,
-                 const std::vector<uint16_t>& code,
+                 const std::vector<uint8_t>& code,
                  const size_t frame_size_in_bytes,
                  const uint32_t core_spill_mask,
                  const uint32_t fp_spill_mask,
