+++ src/compiler/CompilerIR.h
@@ -293,7 +293,7 @@ struct CompilationUnit {
     int totalSize;                      // header + code size
     AssemblerStatus assemblerStatus;    // Success or fix and retry
     int assemblerRetries;
-    std::vector<uint16_t> codeBuffer;
+    std::vector<uint8_t> codeBuffer;
     std::vector<uint32_t> mappingTable;
     std::vector<uint16_t> coreVmapTable;
     std::vector<uint16_t> fpVmapTable;
